refactor: split src/utils/display.js into specialized submodules
Fragmented the massive src/utils/display.js file into specialized submodules
within src/utils/display/ to improve maintainability and testability.

- Extracted formatters, header rendering, Solomon ruling logic,
  session summary, and event handlers into dedicated files.
- Removed the barrel file (src/utils/display.js) in favor of direct imports.
- Updated all code and tests to use direct imports from submodules.
- Verified that all 3130 tests pass correctly.
